move all source files into lib directory to clean thigns up the dist folder really got quite messy
