+++ src/index.ts
@@ -1,8 +1,8 @@
-export { twMerge } from './tailwind-merge'
-export { getDefaultConfig } from './default-config'
-export { extendTailwindMerge } from './extend-tailwind-merge'
-export { createTailwindMerge } from './create-tailwind-merge'
-export type { Config } from './types'
-export * as validators from './validators'
-export { mergeConfigs } from './merge-configs'
-export { fromTheme } from './from-theme'
+export { twMerge } from './lib/tailwind-merge'
+export { getDefaultConfig } from './lib/default-config'
+export { extendTailwindMerge } from './lib/extend-tailwind-merge'
+export { createTailwindMerge } from './lib/create-tailwind-merge'
+export type { Config } from './lib/types'
+export * as validators from './lib/validators'
+export { mergeConfigs } from './lib/merge-configs'
+export { fromTheme } from './lib/from-theme'
