+++ src/lib/default-config.ts
