+++ src/lib/merge-configs.ts
